created LoadVideoFrame class to share load_vid_frame() across generators
